# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xecurity Platform",
    "description": "Xecurity Platform Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/deployments/deployments/": {
      "post": {
        "tags": [
          "deployments",
          "deployments"
        ],
        "summary": "Create Deployment",
        "operationId": "create_deployment_api_v1_deployments_deployments__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "deployments",
          "deployments"
        ],
        "summary": "List Deployments",
        "operationId": "list_deployments_api_v1_deployments_deployments__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeploymentResponse"
                  },
                  "title": "Response List Deployments Api V1 Deployments Deployments  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deployments/deployments/{deployment_id}": {
      "get": {
        "tags": [
          "deployments",
          "deployments"
        ],
        "summary": "Get Deployment",
        "operationId": "get_deployment_api_v1_deployments_deployments__deployment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deployment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "deployments",
          "deployments"
        ],
        "summary": "Update Deployment",
        "operationId": "update_deployment_api_v1_deployments_deployments__deployment_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deployment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "deployments",
          "deployments"
        ],
        "summary": "Delete Deployment",
        "operationId": "delete_deployment_api_v1_deployments_deployments__deployment_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deployment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/licenses/{deployment_id}/issue": {
      "post": {
        "tags": [
          "licenses"
        ],
        "summary": "Issue License",
        "description": "Issue a new license for a deployment from a subscription.",
        "operationId": "issue_license_api_v1_licenses__deployment_id__issue_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "deployment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Deployment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseIssue"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/licenses/{license_id}/revoke": {
      "post": {
        "tags": [
          "licenses"
        ],
        "summary": "Revoke License",
        "description": "Revoke an active license.",
        "operationId": "revoke_license_api_v1_licenses__license_id__revoke_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "license_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "License Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseRevoke"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/licenses/{license_id}": {
      "get": {
        "tags": [
          "licenses"
        ],
        "summary": "Get License",
        "description": "Get a license by ID.",
        "operationId": "get_license_api_v1_licenses__license_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "license_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "License Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/licenses/": {
      "get": {
        "tags": [
          "licenses"
        ],
        "summary": "List Licenses",
        "description": "List licenses.",
        "operationId": "list_licenses_api_v1_licenses__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LicenseResponse"
                  },
                  "title": "Response List Licenses Api V1 Licenses  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeploymentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "environment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeploymentEnvironment"
              },
              {
                "type": "null"
              }
            ],
            "default": "PRODUCTION"
          },
          "current_release_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Release Version"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "organization_id",
          "product_id"
        ],
        "title": "DeploymentCreate"
      },
      "DeploymentEnvironment": {
        "type": "string",
        "enum": [
          "DEVELOPMENT",
          "STAGING",
          "PRODUCTION"
        ],
        "title": "DeploymentEnvironment"
      },
      "DeploymentResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "environment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeploymentEnvironment"
              },
              {
                "type": "null"
              }
            ],
            "default": "PRODUCTION"
          },
          "current_release_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Release Version"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "status": {
            "$ref": "#/components/schemas/DeploymentStatus"
          },
          "last_ping_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ping At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "organization_id",
          "product_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DeploymentResponse"
      },
      "DeploymentStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "INACTIVE",
          "DECOMMISSIONED"
        ],
        "title": "DeploymentStatus"
      },
      "DeploymentUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "environment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeploymentEnvironment"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeploymentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_release_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Release Version"
          }
        },
        "type": "object",
        "title": "DeploymentUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LicenseIssue": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          }
        },
        "type": "object",
        "required": [
          "subscription_id"
        ],
        "title": "LicenseIssue"
      },
      "LicenseResponse": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "deployment_id": {
            "type": "string",
            "title": "Deployment Id"
          },
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "product_plan_id": {
            "type": "string",
            "title": "Product Plan Id"
          },
          "status": {
            "$ref": "#/components/schemas/LicenseStatus"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "payload_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload Json"
          }
        },
        "type": "object",
        "required": [
          "id",
          "deployment_id",
          "subscription_id",
          "organization_id",
          "product_id",
          "product_plan_id",
          "status",
          "issued_at",
          "payload_json"
        ],
        "title": "LicenseResponse"
      },
      "LicenseRevoke": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "LicenseRevoke"
      },
      "LicenseStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "REVOKED",
          "EXPIRED"
        ],
        "title": "LicenseStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}